refactor: passive as data
