+++ simaple/character/passive_skill.py
