+++ simaple/data/passive/resources/default_active_skill/archmagefb.yaml
