+++ simaple/data/passive/resources/default_active_skill/archmagetc.yaml
