+++ simaple/data/passive/resources/default_active_skill/bishop.yaml
