+++ simaple/data/passive/resources/passive_skill/archmagefb.yaml
