+++ simaple/data/passive/resources/passive_skill/archmagetc.yaml
