+++ simaple/data/passive/resources/passive_skill/bishop.yaml
